+++ opensquirrel/passes/mapper/__init__.py
@@ -1,4 +1,5 @@
 from opensquirrel.passes.mapper.general_mapper import Mapper
+from opensquirrel.passes.mapper.mip_mapper import MIPMapper
 from opensquirrel.passes.mapper.simple_mappers import HardcodedMapper, IdentityMapper, RandomMapper
 
-__all__ = ["HardcodedMapper", "IdentityMapper", "Mapper", "RandomMapper"]
+__all__ = ["HardcodedMapper", "IdentityMapper", "MIPMapper", "Mapper", "RandomMapper"]