+++ CHANGELOG.md
@@ -10,11 +10,21 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * **Fixed** for any bug fixes.
 * **Removed** for now removed features.
 
+## [ M.m.P ] - [ xxxx-yy-zz ]
+
+### Added
+
+- `MIPMapper` mapper pass
+
 ## [ 0.5.0 ] - [ 2025-05-28 ]
 
 ### Added
 
-- `asm_filter` method to the `Circuit` class to filter-out assembly declarations by backend name
+- `asm_filter` method to the `Circuit` class to filter out assembly declarations by backend name
+
+### Fixed
+
+- `RoutingValidator` ignores assembly declarations
 
 ### Changed